@@ -1,13 +1,11 @@
-import {ApplicationConfig, provideZoneChangeDetection} from "@angular/core";
-import {HttpClientModule, HttpClient, provideHttpClient} from "@angular/common/http";
+import { ApplicationConfig, provideBrowserGlobalErrorListeners, provideZoneChangeDetection } from '@angular/core';
+import { HttpClient, provideHttpClient} from "@angular/common/http";
 import {provideTranslateService, TranslateLoader} from "@ngx-translate/core";
 import {TranslateHttpLoader} from '@ngx-translate/http-loader';
 import { provideRouter } from '@angular/router';
 import { routes } from './app.routes';
-
 const httpLoaderFactory: (http: HttpClient) => TranslateHttpLoader = (http: HttpClient) =>
-    new TranslateHttpLoader(http, './assets/i18n/', '.json');
-
+    new TranslateHttpLoader();
 export const appConfig: ApplicationConfig = {
   providers: [
     provideRouter(routes),
